Move the start/continue trace tests to files based on message operation (send/publish)
